server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.66.5
python-dotenv==1.0.0
numpy==1.24.3
soundfile==0.12.1
